python/requirements.txt
# Core dependencies for AutoMCAgent Python orchestrator

# WebSocket for bridge communication
websockets>=12.0

# Configuration management
pyyaml>=6.0

# Async support
aiohttp>=3.9.0

# Logging enhancements
colorlog>=6.8.0

# Testing (optional but recommended)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Future: LLM integration (Phase 5)
# ollama-python>=0.1.0
# langchain>=0.1.0
